fix: voteSignBytes sfixed64 encoding and add cross-validation tests
- Fix height/round encoding: varint → sfixed64 (8-byte little-endian)
- Add varint length prefix (MarshalDelimited for Amino compatibility)
- Add encodeSfixed64 helper function
- Remove old encodeTimestamp(uint256), use encodeTimestamp(uint128 nanos)
- Add cross-validation tests for encodeTimestamp (nanos) and voteSignBytes

+++ contracts/utils/Header.sol
@@ -40,7 +40,7 @@ library Header {
         headerBytes[2] = Encode.cdcEncodeInt64(uint256(header.height));
 
         // Field 3: Time (StdTimeMarshal → Timestamp)
-        headerBytes[3] = Encode.encodeTimestamp(uint256(header.time));
+        headerBytes[3] = Encode.encodeTimestamp(header.time);
 
         // Field 4: LastBlockId (proto.Marshal, empty if not present)
         if (header.hasLastBlockId) {
